requirements/default.txt
feedparser
celery
django-celery
BeautifulSoup
httplib2
html5lib
pytz
fake-useragent
nltk==2.0.4